Refactor: complete analyzer overhaul addressing all 8 critical issues
1. RSI now uses Wilder's smoothing (industry standard)
   - Matches TradingView, MT4, all major platforms
   - Old simple-average RSI diverged from real chart values

2. MACD is now O(n) single-pass via MACDSeries()
   - Old version was O(n²) - catastrophic in backtests
   - New version computes EMA history in one loop

3. Confidence driven by calibrated win rate (ScoreTierMap)
   - Falls back to formula only when < 30 trades in tier
   - CalibrateFromBacktest() populates ScoreTierMap from history
   - Confidence now reflects actual historical performance

4. ADX added - gates ranging markets before any scoring
   - ADX < 15: skip (ranging market, no reliable signals)
   - ADX > 25: directional weight added (+DI vs -DI)
   - Eliminates the #1 killer of short-expiry signal bots

5. Weighted scoring replaces integer votes
   - RSI extreme: 2.5, RSI: 1.5
   - 15m MTF: 2.0, 5m MTF: 1.5 (higher TF = more authority)
   - EMA alignment: 0.5 (noisy on 1m)
   - Volume: 0.3 (confirms, not predicts)
   - Threshold is now float (default: 3.5)

6. Stochastic oscillator added
   - Confirms RSI oversold/overbought signals
   - Stoch crossover adds directional weight

7. ComputeWeightedScore: single source of truth
   - AnalyzeAsset and BacktestAsset both call same function
   - No more diverging logic between live and backtest
   - ScoreInput/ScoreOutput structs for clean API

8. ScoreTierStats + CalibrateFromBacktest
   - Backtester now tracks win rate per score tier
   - Live analyzer uses these stats for accurate confidence

+++ cmd/signal-generator/main.go
@@ -68,7 +68,7 @@ func main() {
 	// Create analyzer with config-driven settings
 	analyzerCfg := analyzer.DefaultConfig()
 	if cfg.Analyzer.SignalThreshold > 0 {
-		analyzerCfg.SignalThreshold = cfg.Analyzer.SignalThreshold
+		analyzerCfg.SignalThreshold = float64(cfg.Analyzer.SignalThreshold)
 	}
 	if cfg.Analyzer.SignalCooldown > 0 {
 		analyzerCfg.SignalCooldown = cfg.Analyzer.SignalCooldown
@@ -106,7 +106,7 @@ func main() {
 	logger.Info().
 		Strs("assets", validAssets).
 		Int("interval_sec", intervalSec).
-		Int("signal_threshold", analyzerCfg.SignalThreshold).
+		Float64("signal_threshold", analyzerCfg.SignalThreshold).
 		Int("signal_cooldown_min", analyzerCfg.SignalCooldown).
 		Msg("✓ Signal generator ready")
